Replace UIPageViewController with UIScrollView
